Correction on "Type Entity" and Im[plemented CurD on Company

+++ src/main/java/com/example/service/CompanyDeleteService.java
@@ -1,7 +1,7 @@
 package com.example.service;
 
 public interface CompanyDeleteService {
-	void deleteById(int id);
+	void deleteById(Long id);
 
 
 }
