+++ src/main/java/com/example/service/CompanyUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Company;
 
 public interface CompanyUpdateService {
-	Company updateData(Company company , int id);
+	Company updateData(Company company , Long id);
 
 }
